Fix MetaBinarySearch index-0 bug and InterpolationSearch integer overflow
- MetaBinarySearch: add post-loop check for arr[pos]==key so elements at
  index 0 are correctly found
- InterpolationSearch: cast to long before multiplication to prevent
  integer overflow with large values

+++ src/main/java/com/practice/dsalgo/searching/InterpolationSearch.java
@@ -121,7 +121,7 @@ public class InterpolationSearch {
 			}
 
 			// Estimate the position using interpolation formula
-			int pos = low + ((key - arr[low]) * (high - low)) / (arr[high] - arr[low]);
+			int pos = low + (int)(((long)(key - arr[low]) * (high - low)) / (arr[high] - arr[low]));
 
 			if (arr[pos] == key) {
 				return pos;
